fix : better formating of source code & improve design for code-editor
The first part was the most tricky (better formating of source code).

At first, lines were separated by comma ',' internally. This meant that whenever a comma was encountered, the code editor will replace it with a new line. For instance code like 'post, err := db.getPost()' was displayed in 2 lines rather than 1.

Turned out that the issue was in the frontend. First, I used the wrong codemirror function 'state.text.toStriong(). This return a string representing the code line by line, seperated by comma'. Now I am using 'state.toString()'.

However this did not solve the issue yet. I have noticed as well that all '\n' were deleted when they arrived to the server. After so much looking around, I did realise that the issue was still in the frontend. You see, I was using a form to send data. Since 'codemirror' data are in JS land, I used an intermediary to interact with the form. It was '<input /> the culprit. Whatever is set into 'input' dom element through the value property will lose all '\n' and '\r' character. It make sense, 'input' are supposed to be on 1 line only. To solve the issue I switched to a '<texterarea>' tag, and all was well

The funny thing was that, this was a very bug that I only decided to tackle now since I am nearing the launch date of the project. If my memory serve me right, this bug was present as early as the first commit